# Avoid unneeded calculation of ROI in MIImageView