--- a/docs/release_notes/next/fix-2030-slow-imageview
+++ b/docs/release_notes/next/fix-2030-slow-imageview
@@ -1,0 +1,1 @@
+#2031 : Speed up imageview by avoiding unneeded calculation