### Issue

Closes #2030

### Description

Previously the ROI line plot was recalculated for every movement to a new slice, even though this does not change the plot and the plot may not be visible.

Now, this is only done on changes to the ROI and only if the ROI plot is visible.

There is a change to only show the ROI info in the message if the ROI is visible. This will need the screenshots updating.

There is also an optimisation to only call `mean()` once.

For me this takes the run time for `timeLineChanged()` with the 2k x 2k x 1k flower dataset from ~0.4 to ~0.01 seconds. 
```
Before:
INFO:         1    0.000    0.000    0.469    0.469 /.../ImageView.py:810(timeLineChanged)
INFO:         1    0.000    0.000    0.454    0.454 /.../ImageView.py:810(timeLineChanged)
INFO:         1    0.000    0.000    0.516    0.516 /.../ImageView.py:810(timeLineChanged)

After:
INFO:         1    0.000    0.000    0.010    0.010 /.../ImageView.py:810(timeLineChanged)
INFO:         1    0.000    0.000    0.006    0.006 /.../ImageView.py:810(timeLineChanged)
INFO:         1    0.000    0.000    0.008    0.008 /.../ImageView.py:810(timeLineChanged)
```

### Testing &  Acceptance Criteria 

Test the responsiveness of moving though a large dataset in the main window. Should be noticeable for a 1k x 1k or 2k x 2k image stack, or benchmarkable (see issue) for a smaller stack.

Check that ROI message and line plot show when clicking the ROI button, and go after unclicking it. Moving through the imagestack should still be fast even with the plot shown, as its only recalculated on ROI moves.

Also check that the ROI selector in the the crop operation works. And the image views in compare images and safe apply work.

### Documentation

Release notes