@samtygier-stfc 2024-02-12
I may have reintroduced #1776 :-(