@JackEAllen 2024-02-14
Actually , just noticed a potential bug. I can't crop the image using operations>crop coordinates. after selecting a new roi shape, the roi returns back to 0,0,200,200