@samtygier-stfc 2024-02-14
Yes, might be a good candidate for a system test